requirements.txt
reduct-py~=1.8.1
aiomqtt~=2.0.0